+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix confusing error message from `assert_items_include` function.
 - Print `(no reason specified)` message instead of `nil` value when the test is 
   skipped and no reason is specified.
+- Check `net_box_uri` param is less than max Unix domain socket path length.
 
 ## 0.5.6